The most significant changes are seen in the `Alrorythms and data structure 2.vcxproj` and `Alrorythms and data structure 2.vcxproj.filters` files, `header.h`, and `main.cpp`.
In the `Alrorythms and data structure 2.vcxproj` and `Alrorythms and data structure 2.vcxproj.filters` files, new resources such as `resource.h`, `Alrorythms and data structure 2.rc`, and `icon.ico` have been added.

The `header.h` file has been updated with new functions for deleting airlines and airports, along with additional validation and utility functions. The structure of `Airline` and `Airport` has also been updated with comments and formatting.

In `main.cpp`, the main function has been significantly refactored. The code for setting the locale and console code page has been moved to the top of the function. The main loop has been restructured to use a switch statement for menu selection. The code for each menu option has been updated with additional error checking and user prompts. The code for memory release before exiting has been moved into the case for the exit menu option. The console code page is reset to its original value at the end of the function.

The resource script for a Microsoft Visual C++ application has been added, which includes resources for both Russian and English (United States) languages, and an icon resource specified as "icon.ico". The resources are defined within conditional compilation blocks, which means they will only be included in the build if certain conditions are met.

In the `resource.h` file, several lines of code have been added, including a definition for `IDI_ICON1` with a value of `101` and some default values defined for new objects.

List of changes:

1. Addition of new resources in `Alrorythms and data structure 2.vcxproj` and `Alrorythms and data structure 2.vcxproj.filters` files.
2. Update of `header.h` with new functions and structure updates.
3. Refactoring of main function in `main.cpp`.
4. Addition of a resource script for a Microsoft Visual C++ application.
5. Addition of several lines of code in `resource.h`.