Added new airline, improved UI prompts, and implemented deleteAirlineByIndex function
In this commit, several changes have been made to improve the user interface and functionality of the program. The airline "Pobeda" has been added to the list of airlines. The `<vector>` library has been included and a new function `deleteAirlineByIndex` has been declared and implemented. This function is used to delete an airline from an airport by its index.

In the main function, unnecessary includes have been removed and confirmation prompts have been added before and after adding an airport and an airline. Checks have been added to ensure the airport list is not empty before adding an airline and before deleting an airport.

In `source.cpp`, the `addAirline` and `deleteAirlineFromAirport` functions have been updated to use the new `deleteAirlineByIndex` function when multiple instances of an airline are found. Confirmation prompts and checks have been added to the `deleteAirport` and `deleteAirportFromList` functions. The error message for an invalid choice in `menuSelection` has been commented out.